# Serialized attributes