--- a/.rubocop_todo.yml
+++ b/.rubocop_todo.yml
@@ -79,7 +79,6 @@ Lint/UnusedBlockArgument:
 Lint/UnusedMethodArgument:
   Exclude:
     - 'lib/axlsx/package.rb'
-    - 'lib/axlsx/util/serialized_attributes.rb'
     - 'lib/axlsx/util/validators.rb'
 
 Lint/UselessAssignment:
@@ -94,11 +93,6 @@ Lint/Void:
     - 'lib/axlsx/util/storage.rb'
     - 'lib/axlsx/workbook/worksheet/data_bar.rb'
     - 'lib/axlsx/workbook/worksheet/pivot_table.rb'
-
-# This cop supports safe autocorrection (--autocorrect).
-Performance/BlockGivenWithExplicitBlock:
-  Exclude:
-    - 'lib/axlsx/util/serialized_attributes.rb'
 
 # Configuration parameters: MinSize.
 Performance/CollectionLiteralInLoop:
@@ -369,7 +363,6 @@ Style/Next:
 Style/NonNilCheck:
   Exclude:
     - 'lib/axlsx/drawing/d_lbls.rb'
-    - 'lib/axlsx/util/serialized_attributes.rb'
     - 'lib/axlsx/workbook/worksheet/col.rb'
     - 'lib/axlsx/workbook/worksheet/page_setup.rb'
 
@@ -398,6 +391,7 @@ Style/OptionalBooleanParameter:
   Exclude:
     - 'lib/axlsx.rb'
     - 'lib/axlsx/package.rb'
+    - 'lib/axlsx/util/serialized_attributes.rb'
     - 'lib/axlsx/util/validators.rb'
     - 'lib/axlsx/workbook/workbook.rb'
     - 'lib/axlsx/workbook/worksheet/cell.rb'