--- a/lib/axlsx/stylesheet/num_fmt.rb
+++ b/lib/axlsx/stylesheet/num_fmt.rb
@@ -75,11 +75,7 @@ module Axlsx
 
     # Override to avoid removing underscores
     def serialized_attributes(str = +'', additional_attributes = {})
-      attributes = declared_attributes.merge! additional_attributes
-      attributes.each do |key, value|
-        str << "#{Axlsx.camel(key, false)}=\"#{Axlsx.booleanize(value)}\" "
-      end
-      str
+      super(str, additional_attributes, false)
     end
   end
 end