--- a/lib/axlsx/util/serialized_attributes.rb
+++ b/lib/axlsx/util/serialized_attributes.rb
@@ -41,7 +41,7 @@ module Axlsx
 
     # creates a XML tag with serialized attributes
     # @see SerializedAttributes#serialized_attributes
-    def serialized_tag(tagname, str, additional_attributes = {}, &block)
+    def serialized_tag(tagname, str, additional_attributes = {})
       str << '<' << tagname << ' '
       serialized_attributes(str, additional_attributes)
       if block_given?
@@ -96,7 +96,7 @@ module Axlsx
     # @param [String] str The string instance to which serialized data is appended
     # @param [Array] additional_attributes An array of additional attribute names.
     # @return [String] The serialized output.
-    def serialized_element_attributes(str = +'', additional_attributes = [], &block)
+    def serialized_element_attributes(str = +'', additional_attributes = [])
       attrs = self.class.xml_element_attributes + additional_attributes
       values = Axlsx.instance_values_for(self)
       attrs.each do |attribute_name|