# Description
Serialized attributes are usually declared in snake case as is the convention in Ruby. However, they need to be written in xml in camel case. So the `serialized_attributes` converts each serialied attribute into camel case on each invocation. We can avoid the time / memory for this conversion by caching the conversion the first time and re-using it each subsequent time.

We also avoid using `instance_values_for` to grab all instance variables only to reject ones with nil values and ones that are not serialied attributes. Instead, we can grab only the instance variables that correspond to defined serialied attributes. 

Finally, `Axlsx#booleanize` is commonly used to convert `true` / `false` into `1` / `0` when outputting values to the xml string. However, instead of returning `1` / `0` we can return `'1'` / `'0'` and avoid the `to_s` from needing to do anything.

# Benchmarks
All benchmarks used Ruby 3.2.1 on my macbook pro.

The performance benchmarks varied enough that it is difficult to tell how much of a speed boost we get. I did feel most of the time the new code was a bit faster, but there were instances where the new code was slightly slower. 

Either way, we definitely see a drop in allocated memory.

## Before
```
└─▪ test/benchmark.rb 
...
                                     user     system      total        real
axlsx_noautowidth                0.694793   0.002791   0.697584 (  0.697599)
axlsx_autowidth                  0.678625   0.002049   0.680674 (  0.680835)
axlsx_shared                     0.734707   0.002182   0.736889 (  0.736920)
axlsx_stream                     0.682153   0.001656   0.683809 (  0.683833)
axlsx_zip_command                0.641514   0.020684   0.713513 (  0.714249)
csv                              0.075997   0.005555   0.081552 (  0.081607)

└─▪ test/profile_memory.rb 
Total allocated: 103045567 bytes (1634154 objects)
```

## After
```
└─▪ test/benchmark.rb 
...
                                     user     system      total        real
axlsx_noautowidth                0.709162   0.003540   0.712702 (  0.712935)
axlsx_autowidth                  0.710035   0.002539   0.712574 (  0.712763)
axlsx_shared                     0.721251   0.001731   0.722982 (  0.723049)
axlsx_stream                     0.668538   0.001325   0.669863 (  0.669898)
axlsx_zip_command                0.639397   0.020657   0.712318 (  0.712707)
csv                              0.068726   0.004731   0.073457 (  0.073459)

└─▪ test/profile_memory.rb
Total allocated: 95239839 bytes (1503455 objects)
```

### Your checklist for this pull request
- [x] I have reviewed the [guidelines for contributing](../blob/master/CONTRIBUTING.md) to this repository.
- [x] I have added (or updated) appropriate tests if this PR fixes a bug or adds a feature.
- [x] If this PR doesn't need tests (docs change), I added `[ci skip]` to the title of the PR.
- [x] If this closes any issues, I have added "Closes `#issue`" to the PR description or my commit messages.
- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.
- [ ] I added an entry to the [changelog](../blob/master/CHANGELOG.md).